+++ src/commands/block.rs
@@ -1,15 +1,7 @@
 //! Block Device related commands for Neotron OS
 
+use super::{parse_u64, parse_u8};
 use crate::{bios, osprint, osprintln, Ctx, API};
-
-pub static LSBLK_ITEM: menu::Item<Ctx> = menu::Item {
-    item_type: menu::ItemType::Callback {
-        function: lsblk,
-        parameters: &[],
-    },
-    command: "lsblk",
-    help: Some("List all the Block Devices"),
-};
 
 pub static READ_ITEM: menu::Item<Ctx> = menu::Item {
     item_type: menu::ItemType::Callback {
@@ -29,66 +21,21 @@ pub static READ_ITEM: menu::Item<Ctx> = menu::Item {
     help: Some("Display one disk block, as hex"),
 };
 
-/// Called when the "lsblk" command is executed.
-fn lsblk(_menu: &menu::Menu<Ctx>, _item: &menu::Item<Ctx>, _args: &[&str], _ctx: &mut Ctx) {
-    let api = API.get();
-    let mut found = false;
-
-    osprintln!("Block Devices:");
-    for dev_idx in 0..=255u8 {
-        if let bios::FfiOption::Some(device_info) = (api.block_dev_get_info)(dev_idx) {
-            let (bsize, bunits, dsize, dunits) =
-                match device_info.num_blocks * u64::from(device_info.block_size) {
-                    x if x < (1024 * 1024 * 1024) => {
-                        // Under 1 GiB, give it in 10s of MiB
-                        (10 * x / (1024 * 1024), "MiB", x / 100_000, "MB")
-                    }
-                    x => {
-                        // Anything else in GiB
-                        (10 * x / (1024 * 1024 * 1024), "GiB", x / 100_000_000, "GB")
-                    }
-                };
-            osprintln!("Device {}:", dev_idx);
-            osprintln!("          Name: {}", device_info.name);
-            osprintln!("          Type: {:?}", device_info.device_type);
-            osprintln!("    Block size: {}", device_info.block_size);
-            osprintln!("    Num Blocks: {}", device_info.num_blocks);
-            osprintln!(
-                "     Card Size: {}.{} {} ({}.{} {})",
-                bsize / 10,
-                bsize % 10,
-                bunits,
-                dsize / 10,
-                dsize % 10,
-                dunits
-            );
-            osprintln!("     Ejectable: {}", device_info.ejectable);
-            osprintln!("     Removable: {}", device_info.removable);
-            osprintln!(" Media Present: {}", device_info.media_present);
-            osprintln!("     Read Only: {}", device_info.read_only);
-            found = true;
-        }
-    }
-    if !found {
-        osprintln!("  None");
-    }
-}
-
 /// Called when the "read_block" command is executed.
 fn read_block(_menu: &menu::Menu<Ctx>, _item: &menu::Item<Ctx>, args: &[&str], _ctx: &mut Ctx) {
     let api = API.get();
-    let Ok(dev_idx) = args[0].parse::<u8>() else {
+    let Ok(device_idx) = parse_u8(args[0]) else {
         osprintln!("Couldn't parse {:?}", args[0]);
         return;
     };
-    let Ok(block_idx) = args[1].parse::<u64>() else {
+    let Ok(block_idx) = parse_u64(args[1]) else {
         osprintln!("Couldn't parse {:?}", args[1]);
         return;
     };
     osprintln!("Reading block {}:", block_idx);
     let mut buffer = [0u8; 512];
     match (api.block_read)(
-        dev_idx,
+        device_idx,
         bios::block_dev::BlockIdx(block_idx),
         1,
         bios::FfiBuffer::new(&mut buffer),
@@ -110,3 +57,5 @@ fn read_block(_menu: &menu::Menu<Ctx>, _item: &menu::Item<Ctx>, args: &[&str], _
         }
     }
 }
+
+// End of file