+++ src/commands/input.rs
@@ -42,3 +42,5 @@ fn kbtest(_menu: &menu::Menu<Ctx>, _item: &menu::Item<Ctx>, _args: &[&str], _ctx
     }
     osprintln!("Finished.");
 }
+
+// End of file