+++ src/fs.rs
@@ -80,3 +80,5 @@ impl embedded_sdmmc::TimeSource for BiosTime {
         }
     }
 }
+
+// End of file