+++ src/program.rs
@@ -177,12 +177,7 @@ impl TransientProgramArea {
 
     /// Borrow the TPA region as a slice of words
     pub fn as_slice_u32(&mut self) -> &mut [u32] {
-        unsafe {
-            core::slice::from_raw_parts_mut(
-                self.memory_bottom,
-                self.memory_top.offset_from(self.memory_bottom) as usize,
-            )
-        }
+        unsafe { core::slice::from_raw_parts_mut(self.memory_bottom, self.size_words()) }
     }
 
     /// Borrow the TPA region as a slice of bytes
@@ -190,10 +185,14 @@ impl TransientProgramArea {
         unsafe {
             core::slice::from_raw_parts_mut(
                 self.memory_bottom as *mut u8,
-                (self.memory_top.offset_from(self.memory_bottom) as usize)
-                    * core::mem::size_of::<u32>(),
+                self.size_words() * core::mem::size_of::<u32>(),
             )
         }
+    }
+
+    /// Size of the TPA in 32-bit words
+    fn size_words(&self) -> usize {
+        unsafe { self.memory_top.offset_from(self.memory_bottom) as usize }
     }
 
     /// Loads a program from disk into the Transient Program Area.
@@ -269,6 +268,38 @@ impl TransientProgramArea {
 
         self.last_entry = 0;
         Ok(result)
+    }
+
+    /// Move data to the top of TPA and make TPA shorter.
+    ///
+    /// Moves `size` bytes to the top of the TPA, and then pretends the TPA is
+    /// `size` bytes shorter than it was.
+    ///
+    /// `size` will be rounded up to a multiple of 4.
+    ///
+    /// Panics if `n` is too big to fit in the TPA.
+    ///
+    /// Returns a pointer to the data that now sits outside of the TPA. There
+    /// will be `size` bytes at this address but you must manage the lifetimes
+    /// yourself.
+    pub fn steal_top(&mut self, size: usize) -> *const u8 {
+        let stolen_words = (size + 3) / 4;
+        if stolen_words >= self.size_words() {
+            panic!("Stole too much from TPA!");
+        }
+        unsafe {
+            // Top goes down to free memory above it
+            let new_top = self.memory_top.sub(stolen_words);
+            // Copy the data from the bottom to above the newly reduced TPA
+            core::ptr::copy(self.memory_bottom, new_top, stolen_words);
+            new_top as *mut u8
+        }
+    }
+
+    /// Restore the TPA back where it was.
+    pub unsafe fn restore_top(&mut self, size: usize) {
+        let restored_words = (size + 3) / 4;
+        self.memory_top = self.memory_top.add(restored_words);
     }
 }